Wrap move name families with more than 1 letter with `<>` angle brackets.
`alg.js` currently has a runtime `allowLongMoveNames(true)` config option. This
commit updates `getmovename()` to emit all moves in a format that `alg.js` will
accept when that option is enabled.

We may be able to get rid of the angle brackets in the long term, but in the
short term this allows long move names with an alg parser that:

1) is puzzle-agnostic, and
2) allow leaving out whitespace between moves.